fixed mutation

+++ src/main/scala/queries/CityQueries.scala
@@ -6,7 +6,7 @@ import doobie.implicits.toSqlInterpolator
 object CityQueries {
   private val insertFr: Fragment = fr"INSERT INTO cities (name, country_id, is_capital)"
 
-  def createCity(name: String, countryId: Long, isCapital: Boolean): ConnectionIO[Long] =
+  def createCity(name: String, countryId: Int, isCapital: Boolean): ConnectionIO[Int] =
     (insertFr ++ sql"VALUES ($name, $countryId, $isCapital)").update
-      .withUniqueGeneratedKeys[Long]("id")
+      .withUniqueGeneratedKeys[Int]("id")
 }
